requirements.txt
dnspython>=2.4.0
colorama>=0.4.6